Connect the game and view with handshake events
The game has a new simple event that alerts when the game state changes.
The view also has a new signal event that alerts when it is ready for
the new game state.  These were implemented to allow the game and the
view to act independently but stay in sync